Add css and html files, Create skeleton of page
I have created the basic structure of files for the webpage. I have add
index.html file and have added the styles.css.

In the index.html file I have created the basic skeleton of the website
with the basic wording and buttons added already. I have labeled
sections to give me the general direction and gudiance where I want to
go with the website.